// generated/src/app/api-models/trading-partner-relations.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class TradingPartnerRelations {
  seqTpRelations: number;
  receivingTpId: string;
  sourceTpId: string;
  companyCode: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}